requirements.txt
Django==2.0.2
Pillow==8.2.0
psycopg2-binary==2.8.6
